+++ CHANGELOG.md
@@ -20,6 +20,10 @@
 * Added **run** command which runs a command in the Demisto playground.
 * Added **upload** command which uploads an integration or a script to a Demisto instance.
 * Fixed and issue where **validate** checked if release notes exist for new integrations and scripts.
+* Added **generate-test-playbook** command which generates a basic test playbook for an integration or a script.
+* **validate** now supports indicator fields.
+* Fixed an issue with layouts scheme validation.
+* Adding **init** command.
 * Added **json-to-outputs** command which generates the yaml section for outputs from an API raw response.
 
 ### 0.2.6